修改code报警

+++ ressched/services/resschedmgr/resschedfwk/include/plugin_mgr.h
@@ -58,7 +58,7 @@ struct PluginStat {
                 totalTime = (uint32_t)costTime;
                 useCount = 1;
             } else {
-                totalTime += costTime;
+                totalTime += (uint32_t)costTime;
                 useCount += 1;
             }
         }
